# contracts/agent-v2.openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shadow Travel typed Agent API",
    "version": "2.0.0",
    "description": "Travel-only review; owner browser confirmation required."
  },
  "servers": [
    {
      "url": "https://travel.example.com"
    }
  ],
  "paths": {
    "/api/machine/v1/agent/v2/summary": {
      "get": {
        "tags": [
          "agent-v2"
        ],
        "summary": "Summary",
        "operationId": "get_agent_trip_summary",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Machine identity required"
          },
          "403": {
            "description": "Scope, grant or confirmation denied"
          },
          "404": {
            "description": "Resource not authorized or missing"
          },
          "409": {
            "description": "Version or idempotency conflict"
          }
        },
        "security": [
          {
            "AgentBearer": []
          }
        ]
      }
    },
    "/api/machine/v1/agent/v2/trips": {
      "get": {
        "tags": [
          "agent-v2"
        ],
        "summary": "List Trips",
        "operationId": "list_agent_trips",
        "parameters": [
          {
            "name": "grant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Grant Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Machine identity required"
          },
          "403": {
            "description": "Scope, grant or confirmation denied"
          },
          "404": {
            "description": "Resource not authorized or missing"
          },
          "409": {
            "description": "Version or idempotency conflict"
          }
        },
        "security": [
          {
            "AgentBearer": []
          }
        ]
      }
    },
    "/api/machine/v1/agent/v2/trips/{trip_id}": {
      "get": {
        "tags": [
          "agent-v2"
        ],
        "summary": "Get Trip",
        "operationId": "get_agent_trip",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "grant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Grant Id"
            }
          },
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Machine identity required"
          },
          "403": {
            "description": "Scope, grant or confirmation denied"
          },
          "404": {
            "description": "Resource not authorized or missing"
          },
          "409": {
            "description": "Version or idempotency conflict"
          }
        },
        "security": [
          {
            "AgentBearer": []
          }
        ]
      }
    },
    "/api/machine/v1/agent/v2/trips/{trip_id}/reservations": {
      "get": {
        "tags": [
          "agent-v2"
        ],
        "summary": "List Reservations",
        "operationId": "list_agent_reservations",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "grant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Grant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Machine identity required"
          },
          "403": {
            "description": "Scope, grant or confirmation denied"
          },
          "404": {
            "description": "Resource not authorized or missing"
          },
          "409": {
            "description": "Version or idempotency conflict"
          }
        },
        "security": [
          {
            "AgentBearer": []
          }
        ]
      }
    },
    "/api/machine/v1/agent/v2/trips/{trip_id}/readiness": {
      "get": {
        "tags": [
          "agent-v2"
        ],
        "summary": "Readiness",
        "operationId": "get_agent_trip_readiness",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "grant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Grant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Machine identity required"
          },
          "403": {
            "description": "Scope, grant or confirmation denied"
          },
          "404": {
            "description": "Resource not authorized or missing"
          },
          "409": {
            "description": "Version or idempotency conflict"
          }
        },
        "security": [
          {
            "AgentBearer": []
          }
        ]
      }
    },
    "/api/machine/v1/agent/v2/proposals/check": {
      "post": {
        "tags": [
          "agent-v2"
        ],
        "summary": "Check Proposal",
        "operationId": "check_agent_proposal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Proposal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Machine identity required"
          },
          "403": {
            "description": "Scope, grant or confirmation denied"
          },
          "404": {
            "description": "Resource not authorized or missing"
          },
          "409": {
            "description": "Version or idempotency conflict"
          }
        },
        "security": [
          {
            "AgentBearer": []
          }
        ]
      }
    },
...
    "/api/machine/v1/agent/v2/reviews/{review_id}": {
      "get": {
        "tags": [
          "agent-v2"
        ],
        "summary": "Get Review",
        "operationId": "get_agent_review",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Machine identity required"
          },
          "403": {
            "description": "Scope, grant or confirmation denied"
          },
          "404": {
            "description": "Resource not authorized or missing"
          },
          "409": {
            "description": "Version or idempotency conflict"
          }
        },
        "security": [
          {
            "AgentBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "AgentBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}